Move fishhook out of pspawn_payload
